[TEST] Fixes test fails from reports
