+++ abagen/tests/test_allen.py
@@ -3,7 +3,6 @@
 Tests for abagen.allen module
 """
 
-import nibabel as nib
 import numpy as np
 import pandas as pd
 import pytest
@@ -109,39 +108,3 @@ def test_get_samples_in_mask(testfiles, atlas):
 
     # providing the cortical mask (atlas) should reduce # of samples returned
     assert len(allexp) > len(cortexp)
-
-
-def test_coerce_atlas_to_dict(testfiles, atlas):
-    img, info = atlas['image'], atlas['info']
-    donors = ['12876', '15496']
-
-    # test providing single atlas file
-    atl, same = allen.coerce_atlas_to_dict(img, donors, info)
-    assert same
-    assert sorted(atl.keys()) == sorted(donors)
-    imgs = list(atl.values())
-    assert all(imgs[0] is a for a in imgs[1:])
-    assert isinstance(imgs[0].atlas_info, pd.DataFrame)
-
-    # test providing pre-loaded atlas file
-    atl, same = allen.coerce_atlas_to_dict(nib.load(img), donors, info)
-    assert same
-    assert sorted(atl.keys()) == sorted(donors)
-    imgs = list(atl.values())
-    assert all(imgs[0] is a for a in imgs[1:])
-    assert isinstance(imgs[0].atlas_info, pd.DataFrame)
-
-    # test providing dictionary for atlas
-    atlas_dict = {d: img for d in donors}
-    atl, same = allen.coerce_atlas_to_dict(atlas_dict, donors)
-    assert not same
-    assert sorted(atl.keys()) == sorted(donors)
-    imgs = list(atl.values())
-    assert not any(imgs[0] is a for a in imgs[1:])
-    assert imgs[0].atlas_info is None
-
-    with pytest.raises(ValueError):
-        allen.coerce_atlas_to_dict(atlas_dict, donors + ['9861'])
-
-    with pytest.raises(TypeError):
-        allen.coerce_atlas_to_dict('notanatlas', donors)
